Fix 32-bit build warning
On 32-bit, size_t is 32-bit, while
ScatteredStreamWriter::written_previously_ is uint64_t.

ScatteredStreamWriter is not able to create packets larger than 256Mb
with a TraceWriter and is still limited to 32-bit lengths with a heap
buffer, so there's no way to overflow this.

+++ src/shared_lib/stream_writer.h
@@ -29,7 +29,7 @@ inline void UpdateStreamWriter(const protozero::ScatteredStreamWriter& sw,
   w->begin = sw.cur_range().begin;
   w->end = sw.cur_range().end;
   w->write_ptr = sw.write_ptr();
-  w->written_previously = sw.written_previously();
+  w->written_previously = static_cast<size_t>(sw.written_previously());
 }
 
 }  // namespace perfetto
